requirements.txt
customtkinter>=5.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openpyxl>=3.1.0
Pillow>=10.0.0
srtools>=0.1.1
pyzbar>=0.1.9
pyinstaller>=6.0.0
pymupdf>=1.24.0
zbar-py>=1.0.4
zxing-cpp>=2.0.0; sys_platform == 'darwin'
zxingcpp>=2.0.0; sys_platform == 'win32'
qrcode>=7.4.2


